requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
bzip2=1.0.6=h470a237_2
ca-certificates=2018.03.07=0
certifi=2018.4.16=py37_0
django=2.0.5=py37hd476221_0
djangorestframework=3.8.2=py_0
libffi=3.2.1=3
libgcc-ng=7.2.0=hdf63c60_3
libstdcxx-ng=7.2.0=hdf63c60_3
ncurses=6.1=hfc679d8_1
openssl=1.0.2o=h14c3975_1
pip=10.0.1=py37_0
python=3.7.0=h5001a0f_0
pytz=2018.5=py37_0
readline=7.0=haf1bffa_1
setuptools=40.0.0=py37_1
sqlite=3.24.0=h2f33b56_0
tk=8.6.8=0
wheel=0.31.1=py37_1
xz=5.2.4=h470a237_1
zlib=1.2.11=h470a237_3
